+++ lib/src/connectionhandling/mqtt_client_mqtt_secure_connection.dart
@@ -10,12 +10,7 @@ part of mqtt_client;
 /// The MQTT secure connection class
 class MqttSecureConnection extends MqttConnection {
   /// Default constructor
-  MqttSecureConnection(
-      this.trustedCertPath,
-      this.privateKeyFilePath,
-      this.certificateChainPath,
-      this.privateKeyFilePassphrase,
-      events.EventBus eventBus)
+  MqttSecureConnection(this.context, events.EventBus eventBus)
       : super(eventBus);
 
   /// Initializes a new instance of the MqttSecureConnection class.
@@ -25,17 +20,8 @@ class MqttSecureConnection extends MqttConnection {
     connect(server, port);
   }
 
-  /// Trusted certificate file path for use in secure working
-  String trustedCertPath;
-
-  /// The certificate chain path for secure working.
-  String certificateChainPath;
-
-  /// Private key file path
-  String privateKeyFilePath;
-
-  /// Private keyfile passphrase
-  String privateKeyFilePassphrase;
+  /// The security context for secure usage
+  SecurityContext context;
 
   /// Connect
   @override
@@ -44,30 +30,6 @@ class MqttSecureConnection extends MqttConnection {
         Completer<MqttClientConnectionStatus>();
     MqttLogger.log('MqttSecureConnection::connect');
     try {
-      // Connect and save the socket.
-      final SecurityContext context = SecurityContext.defaultContext;
-      if (trustedCertPath != null) {
-        MqttLogger.log(
-            'MqttSecureConnection::connect - trusted cert path is $trustedCertPath');
-        context.setTrustedCertificates(trustedCertPath);
-      }
-      if (certificateChainPath != null) {
-        MqttLogger.log(
-            'MqttSecureConnection::connect - certificate chain file path is $certificateChainPath');
-        context.useCertificateChain(certificateChainPath);
-      }
-      if (privateKeyFilePath != null) {
-        MqttLogger.log(
-            'MqttSecureConnection::connect - private key file path is $privateKeyFilePath');
-        if (privateKeyFilePassphrase != null) {
-          MqttLogger.log(
-              'MqttSecureConnection::connect - private key file passphrase is $privateKeyFilePassphrase');
-          context.usePrivateKey(privateKeyFilePath,
-              password: privateKeyFilePassphrase);
-        } else {
-          context.usePrivateKey(privateKeyFilePath);
-        }
-      }
       SecureSocket.connect(server, port).then((SecureSocket socket) {
         MqttLogger.log('MqttSecureConnection::connect - securing socket');
         client = socket;
@@ -90,13 +52,10 @@ class MqttSecureConnection extends MqttConnection {
           'MqttSecureConnection::Handshake exception to the message broker {$server}:{$port}. Error is ${e.toString()}';
       completer.completeError(e);
       throw NoConnectionException(message);
-    } on TlsException catch(e) {
-      /// Trap the certificate already in hash table exception, this is not really an error
-      if (!e.toString().contains('CERT_ALREADY_IN_HASH_TABLE')) {
-        final String message =
-            'MqttSecureConnection::TLS exception raised on secure connection. Error is ${e.toString()}';
-        throw NoConnectionException(message);
-      }
+    } on TlsException catch (e) {
+      final String message =
+          'MqttSecureConnection::TLS exception raised on secure connection. Error is ${e.toString()}';
+      throw NoConnectionException(message);
     }
     return completer.future;
   }