+++ test/issues/issue54.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:mqtt_client/mqtt_client.dart';
 
 Future<int> main() async {
-  final MqttClient client = MqttClient('test.mosquitto.org', '');
+  final MqttClient client = MqttClient('localhost', '');
   client.logging(on: false);
   client.keepAlivePeriod = 60;
   client.onDisconnected = onDisconnected;