+++ lib/src/mqtt_client.dart
@@ -292,7 +292,7 @@ class MqttClient {
       _connectionHandler?.disconnect();
       returnCode = MqttConnectReturnCode.solicited;
     }
-    _publishingManager.published.close();
+    _publishingManager.published?.close();
     _publishingManager = null;
     _subscriptionsManager = null;
     _keepAlive?.stop();