+++ lib/src/connectionhandling/mqtt_client_mqtt_ws2_connection.dart
@@ -81,6 +81,13 @@ class _DetachedSocket extends Stream<List<int>> implements Socket {
   @override
   bool setOption(SocketOption option, bool enabled) =>
       _socket.setOption(option, enabled);
+
+  @override
+  Uint8List getRawOption(RawSocketOption option) =>
+      _socket.getRawOption(option);
+
+  @override
+  void setRawOption(RawSocketOption option) => _socket.setRawOption(option);
 }
 
 /// The MQTT secure connection class