+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 #5.0.0
-Issues 48, 49 and 50, breaking API change for connection state and security context 
+Issues 48, 49 50 and 52, breaking API change for connection state and security context
 hence the major version bump.
 
 #4.0.0